% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_images}
\alias{summarize_images}
\title{Summarize points at image level}
\usage{
summarize_images(points, class_col = "major_category")
}
\arguments{
\item{points}{A tidy point table.}

\item{class_col}{Class/label column to summarize.}
}
\value{
An image-level summary tibble.
}
\description{
Summarize points at image level
}
\examples{
cpc <- system.file("extdata", "HIW_158_W_U-1.cpc", package = "pointcoral")
summarize_images(read_cpce_file(cpc), class_col = "raw_label")
}
