% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_points_on_image}
\alias{plot_points_on_image}
\title{Plot CPCe points on an image}
\usage{
plot_points_on_image(
  image_path,
  points,
  label_col = "ml_class",
  point_size = 8
)
}
\arguments{
\item{image_path}{Path to an image.}

\item{points}{Point rows for that image.}

\item{label_col}{Column to use for text labels.}

\item{point_size}{Point radius in pixels.}
}
\value{
A \code{magick-image} overlay.
}
\description{
Draws point halos and labels on an image and returns a \code{magick-image} object.
}
